Remove curl transport adapter models and private headers form public layer
