+++ sdk/core/azure-core/src/http/curl/curl_connection.hpp
