+++ sdk/core/azure-core/test/ut/curl_session.hpp
@@ -9,11 +9,13 @@
  *
  */
 
+#include <azure/core/http/curl/curl.hpp>
+#include <curl/curl.h>
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 #include <string>
 
-#include <azure/core/http/curl/curl.hpp>
+#include "../../src/http/curl/curl_connection_pool.hpp"
 
 namespace Azure { namespace Core { namespace Test {
